+++ cmd/list-lights/main.go
@@ -49,7 +49,7 @@ func (a *application) Run(ctx context.Context, sentryClient libsentry.Client) er
 		return errors.Wrap(ctx, err, "get lights failed")
 	}
 
-	lights := pkg.Lights(hueLights)
+	lights := pkg.LightList(hueLights)
 	sort.Sort(lights)
 
 	slog.Info("found lights", "count", len(lights))